Add files via upload
fault severity diagnosis is done using domain adaptation techniques 4 models were buit and tuned to obtain optimum parameters. cwru dataset is choosen for this purpose and it is preprocessed in the form of keys ranging from ball,ir,or and normal severity from 7 to 21 inch faults injected in both fe and de.